refactor(datasets): extract read/search/aggregation routes to routes/read.ts (6d)
Move readLines + the geo/values/metric/words/max/min aggregation handlers into
routes/read.ts as typed thin adapters using the dataset/publicBaseUrl/esAbortContext
accessors. Source-typing fixes that remove recurring router overload friction and
let the new file stay suppression-free:
- application-key default middleware: req typed as base Request (uses accessors only)
- res.throttle/throttleEnd declared on global Express.Response (in rate-limiting.ts)
- tiles.geojson2pbf vtPrepared optional; outputs.results2sheet bookType defaults xlsx
- config.cache schema gains optional 'disabled' (used by read handlers)
