+++ api/src/datasets/utils/tiles.ts
@@ -26,7 +26,7 @@ export const xyz2bbox = (x, y, z) => {
   return [tile2long(x, z), tile2lat(y + 1, z), tile2long(x + 1, z), tile2lat(y, z)]
 }
 
-export const geojson2pbf = async (geojson, xyz, vtPrepared) => {
+export const geojson2pbf = async (geojson: any, xyz: any, vtPrepared?: any) => {
   if (!geojson || !geojson.features || !geojson.features.length) return null
   const buf = Buffer.from(await geojson2pbfPiscina.run({ geojson, xyz, vtPrepared }))
   return buf
